Change method that drops the item

+++ src/main/java/twilightforest/block/DryingRackBlock.java
@@ -116,7 +116,13 @@ public class DryingRackBlock extends BaseEntityBlock implements SimpleWaterlogge
 
 	@Override
 	protected void onRemove(BlockState state, Level level, BlockPos pos, BlockState newState, boolean movedByPiston) {
-		Containers.dropContentsOnDestroy(state, newState, level, pos);
+		double x = pos.getX();
+		double y = pos.getY();
+		double z = pos.getZ();
+		BlockEntity blockEntity = level.getBlockEntity(pos);
+		if (!(blockEntity instanceof DryingRackBlockEntity)) {return;}
+		ItemStack item = ((DryingRackBlockEntity) blockEntity).takeTheItem();
+		Containers.dropItemStack(level, x, y, z, item);
 		super.onRemove(state, level, pos, newState, movedByPiston);
 	}
